# backend/internal/httpapi/routes.go
package httpapi

import (
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/saiaj/cooking_app/backend/internal/httpapi/response"
)

func routes(app *App) http.Handler {
	r := chi.NewRouter()

	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(app.requestLogger)
	r.Use(middleware.Recoverer)
	r.Use(middleware.Timeout(30 * time.Second))

	r.Get("/healthz", func(w http.ResponseWriter, _ *http.Request) {
		w.Header().Set("Content-Type", "text/plain; charset=utf-8")
		if _, err := w.Write([]byte("ok\n")); err != nil {
			app.logger.Warn("write failed", "err", err, "path", "/healthz")
		}
	})

	r.Route("/api/v1", func(r chi.Router) {
		r.NotFound(func(w http.ResponseWriter, r *http.Request) {
			app.writeError(w, r, errNotFound())
		})

		r.MethodNotAllowed(func(w http.ResponseWriter, r *http.Request) {
			app.writeError(w, r, errMethodNotAllowed())
		})

		r.Get("/healthz", func(w http.ResponseWriter, _ *http.Request) {
			if err := response.WriteJSON(w, http.StatusOK, struct {
				OK bool `json:"ok"`
			}{OK: true}); err != nil {
				app.logger.Warn("write failed", "err", err, "path", "/api/v1/healthz")
			}
		})

		r.Route("/auth", func(r chi.Router) {
			r.With(app.loginRateLimitMiddleware).Post("/login", app.handle(app.handleLogin))
			r.With(app.authMiddleware).Post("/logout", app.handle(app.handleLogout))
			r.With(app.authMiddleware).Get("/me", app.handle(app.handleMe))
		})

		r.Route("/tokens", func(r chi.Router) {
			r.Use(app.authMiddleware)
			r.Get("/", app.handle(app.handleTokensList))
			r.Post("/", app.handle(app.handleTokensCreate))
			r.Delete("/{id}", app.handle(app.handleTokensDelete))
		})

		r.Route("/users", func(r chi.Router) {
			r.Use(app.authMiddleware)
			r.Get("/", app.handle(app.handleUsersList))
			r.Post("/", app.handle(app.handleUsersCreate))
			r.Put("/{id}/deactivate", app.handle(app.handleUsersDeactivate))
		})

		r.Route("/tags", func(r chi.Router) {
			r.Use(app.authMiddleware)
			r.Get("/", app.handle(app.handleTagsList))
			r.Post("/", app.handle(app.handleTagsCreate))
			r.Put("/{id}", app.handle(app.handleTagsUpdate))
			r.Delete("/{id}", app.handle(app.handleTagsDelete))
		})

		r.Route("/recipe-books", func(r chi.Router) {
			r.Use(app.authMiddleware)
			r.Get("/", app.handle(app.handleRecipeBooksList))
			r.Post("/", app.handle(app.handleRecipeBooksCreate))
			r.Put("/{id}", app.handle(app.handleRecipeBooksUpdate))
			r.Delete("/{id}", app.handle(app.handleRecipeBooksDelete))
		})

		r.Route("/recipes", func(r chi.Router) {
			r.Use(app.authMiddleware)
			r.Get("/", app.handle(app.handleRecipesList))
			r.Get("/{id}", app.handle(app.handleRecipesGet))
			r.Post("/", app.handle(app.handleRecipesCreate))
			r.Put("/{id}", app.handle(app.handleRecipesUpdate))
			r.Delete("/{id}", app.handle(app.handleRecipesDelete))
			r.Put("/{id}/restore", app.handle(app.handleRecipesRestore))
		})
	})

	return r
}